+++ src/Monify.Console.Tests/Classes/Nested/InRecord/OutterForImmutableArrayTests/WhenEqualsWithObjectIsCalled.cs
@@ -34,16 +34,16 @@ public static class WhenEqualsWithObjectIsCalled
     }
 
     [Fact]
-    public static void GivenDifferentTypeThenThrowInvalidCastException()
+    public static void GivenDifferentTypeThenReturnsFalse()
     {
         // Arrange
         OutterForImmutableArray<int>.Inner subject = new(_sampleValue);
         object other = string.Empty;
 
         // Act
-        Action act = () => subject.Equals(other);
+        bool result = subject.Equals(other);
 
         // Assert
-        _ = Should.Throw<InvalidCastException>(act);
+        result.ShouldBeFalse();
     }
 }