+++ src/Monify.Console.Tests/Structs/Passthrough/OuterTests/WhenInequalityOperatorWithOuterIsCalled.cs
@@ -1,0 +1,51 @@
+namespace Monify.Console.Structs.Passthrough.OuterTests;
+
+public static class WhenInequalityOperatorWithOuterIsCalled
+{
+    private const string SampleValue = "Sample";
+    private const string DifferentValue = "Different";
+
+    [Fact]
+    public static void GivenSameValueThenReturnFalse()
+    {
+        // Arrange
+        Outer left = new(SampleValue);
+        Outer right = new(SampleValue);
+
+        // Act
+        bool actual = left != right;
+
+        // Assert
+        actual.ShouldBeFalse();
+    }
+
+    [Fact]
+    public static void GivenEquivalentValuesThenReturnFalse()
+    {
+        // Arrange
+        string leftValue = new(SampleValue.ToCharArray());
+        string rightValue = new(SampleValue.ToCharArray());
+        Outer left = new(leftValue);
+        Outer right = new(rightValue);
+
+        // Act
+        bool actual = left != right;
+
+        // Assert
+        actual.ShouldBeFalse();
+    }
+
+    [Fact]
+    public static void GivenDifferentValuesThenReturnTrue()
+    {
+        // Arrange
+        Outer left = new(SampleValue);
+        Outer right = new(DifferentValue);
+
+        // Act
+        bool actual = left != right;
+
+        // Assert
+        actual.ShouldBeTrue();
+    }
+}